[FEAT] Login API / add Logo image
