+++ plantweb/src/App.js
@@ -2,10 +2,10 @@ import logo from './logo.svg';
 import './App.css';
 
 import {Header} from "./component/header";
-import {Login} from "./page/login";
-import {Main} from "./page/main";
-import {ShowSearchList} from "./page/showSearchList"
-import {ShowDetails} from "./page/showDetails"
+import {Login} from "./page/login/login";
+import {Main} from "./page/main/main";
+import {ShowSearchList} from "./page/showSearchList/showSearchList"
+import {ShowDetails} from "./page/showDetails/showDetails"
 
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
@@ -14,7 +14,7 @@ const Router = () => {
   return (
      <BrowserRouter>
         <Routes>
-           <Route exact path="/main" element ={<Main/>} />
+           <Route exact path="/" element ={<Main/>} />
            <Route exact path="/login" element ={<Login/>} />
            <Route exact path="/searchList" element ={<ShowSearchList/>} />
            <Route exact path="/details" element ={<ShowDetails/>} />
