+++ plantweb/src/page/myPage/myPage.js
@@ -1,0 +1,9 @@
+export const MyPage = ()=>{
+    return(
+        <div>
+            마이페이지 입니다.
+        </div>
+    );
+}
+
+export default MyPage;
