--- plantweb/src/page/mypage.js
@@ -1,9 +1,0 @@
-export const MyPage = ()=>{
-    return(
-        <div>
-            마이페이지 입니다.
-        </div>
-    );
-}
-
-export default MyPage;
