+++ plantweb/src/page/showDetails/showDetails.css
