+++ plantweb/src/page/showDetails/showDetails.js
@@ -1,4 +1,4 @@
-import Header from "../component/header";
+import Header from "../../component/header";
 import "./showDetails.css"
 
 export const ShowDetails=()=>{
